cli: on POSIX systems, use terminfo (and curses)
When building hledger on a non-windows system, and the new "curses"
cabal flag is true (as it is by default), include the terminfo package.
This will be useful for detecting terminal width and perhaps other
things.

Installation on posix is slightly harder now, since it will need the C
curses dev libraries. The curses flag lets posix users skip that
requirement by adding -f-curses (or cabal might even do that
automatically).