require nightly stackage snapshot for hledger-ui
vty and brick are not yet in the LTS snapshots.
We don't want to list them explicitly as extra-deps since that
can't be conditionalised for windows, which doesn't support them.
So, require a nightly snapshot which does include them.
A downside of this is that most people won't have this
particular snapshot installed. But with luck, stack will
reuse bits from their older snapshots and there won't be too much building.