;site: symlink the main README.md, render it as part of the site
The top-level README.md is usually seen when browsing github or the
cloned source - not on the website - but now we render it as part of
the website to facilitate previewing it locally (with make
site-watch).

As a side effect of symlinking it into site/, it might override site/'s
README when browsing that directory on github.

[ci skip]