;shake manuals: workaround to avoid ANSI codes in plaintext help
This seems to be a groff bug, as of 1.22.3 & 1.22.4. It always calls
grotty with no flags, even with -c, so grotty adds ANSI/SGR escape
codes in the -Tascii output.

Now instead of groff, we call the lower level commands directly,
requiring tbl, eqn, troff, grotty to be in PATH and compatible.
Not ideal.

[ci skip]